+++ packages/@runlightyear/cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @runlightyear/cli
 
+## 0.13.0
+
+### Minor Changes
+
+- Added subscribeProps operation
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.13.0
+
 ## 0.12.1
 
 ### Patch Changes