+++ packages/@runlightyear/github/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/github
 
+## 0.7.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.13.0
+
 ## 0.7.4
 
 ### Patch Changes