+++ packages/@runlightyear/slack/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/slack
 
+## 0.4.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.13.0
+
 ## 0.4.13
 
 ### Patch Changes